requirements-planner.txt
# Requirements for the planner module
# Core dependencies
PyYAML>=6.0
numpy>=1.21.0

# Optional T5 integration for missing field prediction
transformers>=4.30.0
torch>=2.0.0

# Development and testing
pytest>=7.0.0